WIP.  cli refactored into core.  core.detect split from core.changedetection to enable chunking
